requirements.txt
fastapi
uvicorn
python-dotenv
psycopg2-binary
SQLAlchemy
alembic
requests
async-exit-stack
async-generator
pytest
sqlalchemy-utils